replaced `sys.maxint` with `.maxsize`
maxint deprecated https://docs.python.org/3.1/whatsnew/3.0.html#integers

"The sys.maxint constant was removed, since there is no longer a limit to the value of integers. However, sys.maxsize can be used as an integer larger than any practical list or string index. It conforms to the implementation’s “natural” integer size and is typically the same as sys.maxint in previous releases on the same platform (assuming the same build options)."